+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+### WIP
+- Internal fix to bring back spawn options
+
 #### 0.5.6 (December 12, 2013)
 - Add specific WebStorm and IntelliJ files to gitignore
 - Stop worrying about absolute paths